added docs and updated test format
