+++ crates/h264/Cargo.toml
@@ -3,12 +3,15 @@ name = "scuffle-h264"
 version = "0.1.1"
 edition = "2024"
 license = "MIT OR Apache-2.0"
-description = "A pure Rust H.264 header decoder."
 repository = "https://github.com/scufflecloud/scuffle"
 authors = ["Scuffle <opensource@scuffle.cloud>"]
 readme = "README.md"
 documentation = "https://docs.rs/scuffle-h264"
+description = "A pure Rust H.264 header decoder."
 keywords = ["h264", "video", "codec"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(coverage_nightly)'] }
 
 [dependencies]
 bytes = "1.5"
